requirements.txt
torch>=1.13.1
torchvision
numpy
opencv-python
Pillow
tqdm
scikit-image
streamlit